feat(core): export `copyTemplate`
