+++ .changeset/four-crabs-divide.md
@@ -1,0 +1,5 @@
+---
+'@monots/core': minor
+---
+
+Export `copyTemplate` which uses template variables to copy a directory to it's destination. It supports templates within files and also the filenames themeselves can be templated.
