+++ packages/monots__cli/package.json
@@ -58,14 +58,11 @@
     "chalk": "^4.1.2",
     "clipanion": "^3.1.0",
     "delay": "^5.0.0",
-    "lodash-es": "^4.17.21",
     "ora": "^6.0.0",
     "read-pkg-up": "^9.0.0",
-    "recursive-copy": "^2.0.13",
     "update-notifier": "^5.1.0"
   },
   "devDependencies": {
-    "@types/lodash-es": "^4.17.5",
     "@types/update-notifier": "^5.1.0",
     "del": "^6.0.0",
     "nanoid": "^3.1.30",
