+++ packages/monots__core/package.json
@@ -46,6 +46,7 @@
     "@swc/helpers": "^0.2.13",
     "@swc/register": "^0.1.7",
     "@types/js-yaml": "^4.0.4",
+    "@types/lodash-es": "^4.17.5",
     "@types/normalize-path": "^3.0.0",
     "@types/npm-packlist": "^1.1.2",
     "@types/object.pick": "^1.3.1",
@@ -63,11 +64,13 @@
     "jest-diff": "^27.3.1",
     "js-yaml": "^4.1.0",
     "load-json-file": "^7.0.0",
+    "lodash-es": "^4.17.21",
     "make-error": "^1.3.6",
     "normalize-path": "^3.0.0",
     "npm-packlist": "^3.0.0",
     "object.pick": "^1.3.0",
     "parse-json": "^5.2.0",
+    "recursive-copy": "^2.0.13",
     "rollup": "^2.59.0",
     "sort-keys": "^5.0.0",
     "superstruct": "^0.15.3",
